Add building block prefabs and update default map
Added new building block and lobby prefabs to the project, as well as a new emissive material. Updated DefaultEngine.ini to set DebugMap as the default and editor startup map. Enabled additional plugins in the uproject file, including AIAssistant and AnimationWarping, and made changes to several core blueprints and the DebugMap level.